Changed bool* to etl::bitset in LDPC and Convolutional Encoder - Added LDPC Encoder test.

Closes #9 and #28

See merge request acubesat/comms/software/physical_layer!20

+++ inc/PositionRowsEsoteric.hpp
@@ -6,8 +6,7 @@ static constexpr uint8_t generatorRowsEsoteric = 32;
 
 static constexpr uint8_t generatorColumnsEsoteric = 8;
 
-static constexpr int totalQuasiCyclicMatricesEsoteric = generatorColumnsEsoteric * generatorRowsEsoteric;
-
+static constexpr uint16_t totalQuasiCyclicMatricesEsoteric = generatorColumnsEsoteric * generatorRowsEsoteric;
 
 inline static const uint16_t positionRowsEsoteric[totalQuasiCyclicMatricesEsoteric] = {
     0,     63,    128,   194,   266,   336,   388,   444,   508,   572,   639,   708,   764,
